# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Prasso Invenbin API",
        "description": "API endpoints for Prasso Invenbin",
        "contact": {
            "email": "bcp@faxt.com"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/product-statuses": {
            "get": {
                "tags": [
                    "ProductStatuses"
                ],
                "summary": "Get all statuses that apply to products",
                "description": "Retrieve a list of product status records",
                "operationId": "2dedb544a1d062d4486941cbca08435e",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "id": {
                                                "type": "integer",
                                                "example": 1
                                            },
                                            "status": {
                                                "type": "string",
                                                "example": "Active"
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "ProductStatuses"
                ],
                "summary": "Create a new product status",
                "description": "Create a new product status",
                "operationId": "758cf3b6fdda7f068351fc3bb0e588f9",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ProductStatusInputOutput"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ProductStatusInputOutput"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        },
        "/product-statuses/{id}": {
            "get": {
                "tags": [
                    "ProductStatuses"
                ],
                "summary": "Get a specific product status",
                "description": "Display the specified resource.",
                "operationId": "7697e1d45ee291ee218afe103adabb7d",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id of the product status",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ProductStatusInputOutput"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "ProductStatuses"
                ],
                "summary": "Update a product status",
                "description": "Update a product status",
                "operationId": "ae914034f412b37f2644a63e747ee418",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id of the product status",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ProductStatusInputOutput"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ProductStatusInputOutput"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "ProductStatuses"
                ],
                "summary": "Delete a product status",
                "description": "Remove the specified resource from storage.",
                "operationId": "2d10d1f299441504f304455ae086fe7c",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the product status",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "ProductStatus deleted successfully"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "ProductStatusInputOutput": {
                "title": "ErpProductStatus",
                "description": "Erp Product Status model",
                "properties": {
                    "id": {
                        "description": "The unique identifier for the product status",
                        "type": "integer",
                        "format": "int64",
                        "example": 1
                    },
                    "status": {
                        "description": "The status of the product",
                        "type": "string",
                        "example": "Active"
                    },
                    "updated_by": {
                        "description": "The user who last updated the product status",
                        "type": "string",
                        "example": "John Doe"
                    },
                    "created_at": {
                        "description": "The date and time when the product status was created",
                        "type": "string",
                        "format": "date-time",
                        "readOnly": true
                    },
                    "updated_at": {
                        "description": "The date and time when the product status was last updated",
                        "type": "string",
                        "format": "date-time",
                        "readOnly": true
                    }
                },
                "type": "object"
            },
...
        },
...
        "securitySchemes": {
            "sanctum": {
                "type": "apiKey",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "in": "header"
            }
        }
    }
}